log out handle + student name

+++ Frontend/src/components/FormateurSidebar.jsx
@@ -17,9 +17,10 @@ const Sidebar = () => {
   ];
   
   const handlelogout = () => {
+    const confirmed = window.confirm('Are you sure you want to log out?');
+  if (confirmed) {
     sessionStorage.clear();
-    window.location.href = "/Home"
-  };
+    window.location.href = "/Home"}};
 
   return (
     <aside className="ml-[-100%] fixed z-10 top-0 pb-3 px-6 w-full flex flex-col justify-between h-screen border-r bg-neutral-100 transition duration-300 md:w-4/12 lg:ml-0 lg:w-[25%] xl:w-[20%] 2xl:w-[15%]">
